move simple_async_mock to test_utils
... so that it can be re-used
